+++ src/test/java/de/gematik/demis/service/base/clients/mapping/CodeMappingAutoConfigurationTest.java
@@ -53,7 +53,7 @@ class CodeMappingAutoConfigurationTest {
       props.getClient().setBaseUrl("http://example");
       props.getClient().setContextPath("/");
       props.setConceptMaps(List.of("DiseaseA", "LabA"));
-      return new CodeMappingService(client, props, ReloadableCache::new);
+      return new CodeMappingService(client, props, ReloadableCache::new, false);
     }
   }